requirement.txt
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
fabric==3.2.2
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
waitress==3.0.2
Werkzeug==3.1.5
wrapt==2.0.1
